+++ Element/Type/QueryBuilderAdminType.php
@@ -3,33 +3,18 @@
 namespace Mapbender\QueryBuilderBundle\Element\Type;
 
 use Mapbender\CoreBundle\Element\Type\MapbenderTypeTrait;
+use Mapbender\ManagerBundle\Form\Type\YAMLConfigurationType;
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\Extension\Core\Type\CheckboxType;
-use Symfony\Component\Form\Extension\Core\Type\ChoiceType;
-use Symfony\Component\Form\Extension\Core\Type\TextType;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Contracts\Translation\TranslatorInterface;
 
 class QueryBuilderAdminType extends AbstractType
 {
-    protected $dataStoreNames = array();
-
     use MapbenderTypeTrait;
 
-    private TranslatorInterface $trans;
-
-    #public function __construct(TranslatorInterface $trans)
-    #{
-    #    $this->trans = $trans;
-    #}
-
-    /**
-     * @param mixed[]|null $dataStores
-     */
-    public function __construct(array $dataStores, TranslatorInterface $trans)
+    public function __construct(private TranslatorInterface $trans)
     {
-        $this->trans = $trans;
-        $this->dataStoreNames = array_keys($dataStores);
     }
 
     /**
@@ -37,32 +22,12 @@ class QueryBuilderAdminType extends AbstractType
      */
     public function buildForm(FormBuilderInterface $builder, array $options)
     {
-        $dataStoreSelectValues = array_combine($this->dataStoreNames, $this->dataStoreNames);
-        // @todo: add translatable field labels
-
         $builder
-            ->add('source', ChoiceType::class, array(
-                    'choices' => $dataStoreSelectValues,
+            ->add('configuration', YAMLConfigurationType::class, array(
                     'required' => true,
-                    'label' => 'mb.querybuilder.admin.source'
+                    'label' => 'mb.querybuilder.admin.configuration'
                 )
             )
-            ->add('sqlFieldName', TextType::class, array(
-                'required' => true,
-                'label' => 'mb.querybuilder.admin.sqlFieldName'
-            ))
-            ->add('orderByFieldName', TextType::class, array(
-                'required' => true,
-                'label' => 'mb.querybuilder.admin.orderByFieldName'
-            ))
-            ->add('titleFieldName', TextType::class, array(
-                'required' => true,
-                'label' => 'mb.querybuilder.admin.titleFieldName'
-            ))
-            ->add('connectionFieldName', TextType::class, array(
-                'required' => true,
-                'label' => 'mb.querybuilder.admin.connectionFieldName'
-            ))
             ->add('allowCreate', CheckboxType::class, array(
                 'required' => false,
                 'label' => 'mb.querybuilder.admin.allowCreate'
@@ -70,10 +35,6 @@ class QueryBuilderAdminType extends AbstractType
             ->add('allowEdit', CheckboxType::class, array(
                 'required' => false,
                 'label' => 'mb.querybuilder.admin.allowEdit'
-            ))
-            ->add('allowSave', CheckboxType::class, array(
-                'required' => false,
-                'label' => 'mb.querybuilder.admin.allowSave'
             ))
             ->add('allowRemove', CheckboxType::class, array(
                 'required' => false,
@@ -83,16 +44,19 @@ class QueryBuilderAdminType extends AbstractType
                 'required' => false,
                 'label' => 'mb.querybuilder.admin.allowExecute'
             ))
-            ->add('allowExport', CheckboxType::class, array(
+            ->add('allowFileExport', CheckboxType::class, array(
                 'required' => false,
-                'label' => 'mb.querybuilder.admin.allowExport'
+                'label' => 'mb.querybuilder.admin.allowFileExport'
+            ))
+            ->add('allowHtmlExport', CheckboxType::class, array(
+                'required' => false,
+                'label' => 'mb.querybuilder.admin.allowHtmlExport'
             ))
             ->add('allowSearch', CheckboxType::class, array(
                 'required' => false,
                 'label' => 'mb.querybuilder.admin.allowSearch'
             ))
             // @todo: add field for implemented option 'tableColumns'? (can be set via yaml definition; check functionality)
-            // @todo: add field for implemented option 'allowHtmlExport'? (can be set via yaml definition; check functionality)
         ;
     }
 }